app/pyproject.toml
[project]
name = "11labshack-api"
version = "0.1.0"
description = "FastAPI backend for 11labshack"
authors = [
    {name = "User", email = "user@example.com"},
]
dependencies = [
    "fastapi",
    "uvicorn",
    "supabase",
    "python-dotenv",
    "elevenlabs",
    "pydantic",
    "pydantic-settings",
    "httpx",
    "openai",
    "anthropic",
    "google-generativeai"
]
requires-python = ">=3.9"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["routers", "services"]
